+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog]
   - Add ECP admin claim employment task screen
   - Remove verify notice from claim summary
   - Add a note if DQT active alert flag is set to true
+  - Add important notes to claim summary (DQT active alert)
 - All policies
   - Home Address selection
     - Add home address search screen to claim journey